feat(Code/OBJ): Add !@ token.
From @fosler

I have made a minor change to the OBJ language by defining a special token '!@'
that, when used in a context that can access a class's static environment,
evaluates to (an object containing) the environment in which the class was
created -- i.e., in which the token 'class' appears. Because the static
environment of any class starts out extending only the top-level program
environment, the static environment of a class created in an enclosing
environment will not "see" any local variables in the enclosing environment.
Using !@ makes it possible to refer to these local variables.  Here's an
example:

```
define x = 3
define c =
  let
    x = 5
  in
    class end % the static environment of this class is the top-level program environment
<c>x      % => 3
<c><!@>x  % => 5
```

> The last expression can also be written as

```
<<c>!@>x
```

The '!@' feature is implemented by creating a binding of this string (it's not
a variable) to the object that captures the environment of the class in which
the class is created and adding this binding to the static environment of the
class. When used in an expression, '!@' is treated like a variable and its
value is looked up in the current environment. If the current environment does
not contain the static environment of a class, attempting to evaluate '!@'
throws an exception.

---

Closes https://github.com/ourPLCC/course/issues/6